git-svn-id: https://aux1.edublogs.org/repos/incsub/trunk/plugins/coursepress@14674 f53fe46e-8c02-459b-9fdd-93e75771b278

+++ includes/admin-pages/courses.php
@@ -55,7 +55,7 @@ if (isset($_GET['quick_setup'])) {
 
 // Query the courses
     if (isset($_GET['page_num'])) {
-        $page_num = (int)$_GET['page_num'];
+        $page_num = (int) $_GET['page_num'];
     } else {
         $page_num = 1;
     }
@@ -73,6 +73,9 @@ if (isset($_GET['quick_setup'])) {
     }
 
     if (isset($_GET['action']) && $_GET['action'] == 'delete' && isset($_GET['course_id']) && is_numeric($_GET['course_id'])) {
+        if (!isset($_GET['cp_nonce']) || !wp_verify_nonce($_GET['cp_nonce'], 'delete_course')) {
+            die(__('Cheating huh?', 'cp'));
+        }
         $course_object = $course->get_course();
         if (current_user_can('coursepress_delete_course_cap') || (current_user_can('coursepress_delete_my_course_cap') && $course_object->post_author == get_current_user_id())) {
             $course->delete_course($force_delete = true);
@@ -83,13 +86,16 @@ if (isset($_GET['quick_setup'])) {
     }
 
     if (isset($_GET['action']) && $_GET['action'] == 'change_status' && isset($_GET['course_id']) && is_numeric($_GET['course_id'])) {
+        if (!isset($_GET['cp_nonce']) || !wp_verify_nonce($_GET['cp_nonce'], 'change_course_status')) {
+            die(__('Cheating huh?', 'cp'));
+        }
         $course->change_status($_GET['new_status']);
         $message = __('Status for the selected course has been changed successfully.', 'cp');
     }
     ?>
     <div class="wrap nosubsub">
         <div class="icon32" id="icon-themes"><br></div>
-        <h2><?php _e('Courses', 'cp'); ?><?php if (current_user_can('coursepress_create_course_cap')) { ?><a class="add-new-h2" href="admin.php?page=course_details"><?php _e('Add New', 'cp'); ?></a><?php } ?></h2>
+        <h2><?php _e('Courses', 'cp'); ?><?php if (current_user_can('coursepress_create_course_cap')) { ?><a class="add-new-h2" href="<?php echo admin_url('admin.php?page=course_details'); ?>"><?php _e('Add New', 'cp'); ?></a><?php } ?></h2>
 
         <?php
         if (isset($message)) {
@@ -101,7 +107,7 @@ if (isset($_GET['quick_setup'])) {
         <div class="tablenav tablenav-top">
 
             <div class="alignright actions new-actions">
-                <form method="get" action="?page=<?php echo esc_attr($page); ?>" class="search-form">
+                <form method="get" action="<?php echo admin_url('admin.php?page=' . $page); ?>" class="search-form">
                     <p class="search-box">
                         <input type='hidden' name='page' value='<?php echo esc_attr($page); ?>' />
                         <label class="screen-reader-text"><?php _e('Search Courses', 'cp'); ?>:</label>
@@ -111,7 +117,7 @@ if (isset($_GET['quick_setup'])) {
                 </form>
             </div><!--/alignright-->
 
-            <form method="post" action="?page=<?php echo esc_attr($page); ?>" id="posts-filter">
+            <form method="post" action="<?php echo esc_attr(admin_url('admin.php?page=' . $page)); ?>" id="posts-filter">
 
                 <?php if (current_user_can('coursepress_change_course_status_cap') || current_user_can('coursepress_delete_course_cap')) { ?>
                     <div class="alignleft actions">
@@ -183,35 +189,34 @@ if (isset($_GET['quick_setup'])) {
                             $course_object = $course_obj->get_course();
 
                             $style = ( 'alternate' == $style ) ? '' : 'alternate';
-
                             ?>
                             <tr id='user-<?php echo $course_object->ID; ?>' class="<?php echo $style; ?>">
                                 <th scope='row' class='check-column'>
                                     <input type='checkbox' name='courses[]' id='user_<?php echo $course_object->ID; ?>' class='' value='<?php echo $course_object->ID; ?>' />
                                 </th>
-                                <td class="column-course <?php echo $style; ?>"><a href="?page=course_details&course_id=<?php echo $course_object->ID; ?>"><strong><?php echo $course_object->post_title; ?></strong></a><br />
+                                <td class="column-course <?php echo $style; ?>"><a href="<?php echo admin_url('admin.php?page=course_details&course_id=' . $course_object->ID); ?>"><strong><?php echo $course_object->post_title; ?></strong></a><br />
                                     <!-- <div class="course-thumbnail"><img src="<?php echo $course_obj->get_course_thumbnail(); ?>" alt="<?php echo esc_attr($course_object->post_title); ?>" /></div> -->
-                                                      <div class="course_excerpt"><?php echo get_the_course_excerpt($course_object->ID, 55); ?></div>
+                                    <div class="course_excerpt"><?php echo get_the_course_excerpt($course_object->ID, 55); ?></div>
                                     <div class="column-course-units visible-small visible-extra-small">
-                                        <strong><?php _e('Units', 'cp');?>:</strong>
-                                      <?php echo $course_obj->get_units('', 'any', true);?> <?php _e('Units', 'cp');?>,
-                                                      <?php echo $course_obj->get_units('', 'publish', true);?> Published
+                                        <strong><?php _e('Units', 'cp'); ?>:</strong>
+                                        <?php echo $course_obj->get_units('', 'any', true); ?> <?php _e('Units', 'cp'); ?>,
+                                        <?php echo $course_obj->get_units('', 'publish', true); ?> Published
                                     </div>
                                     <div class="column-course-students visible-small visible-extra-small">
-                                        <strong><?php _e('Students', 'cp');?>:</strong>
-                                      <a href="?page=course_details&tab=students&course_id=<?php echo $course_object->ID; ?>"><?php echo $course_obj->get_number_of_students(); ?></a>
+                                        <strong><?php _e('Students', 'cp'); ?>:</strong>
+                                        <a href="<?php echo admin_url('admin.php?page=course_details&tab=students&course_id=' . $course_object->ID); ?>"><?php echo $course_obj->get_number_of_students(); ?></a>
                                     </div>									
                                     <div class="row-actions hide-small hide-extra-small">
-                                        <span class="edit_course"><a href="?page=course_details&course_id=<?php echo $course_object->ID; ?>"><?php _e('Edit', 'cp'); ?></a> | </span>
-                                        <?php if (current_user_can('coursepress_delete_course_cap') || (current_user_can('coursepress_delete_my_course_cap') && $course_object->post_author == get_current_user_id())) { ?>
-                                            <span class="course_units"><a href="?page=course_details&tab=units&course_id=<?php echo $course_object->ID; ?>"><?php _e('Units', 'cp'); ?></a> | </span>
+                                        <span class="edit_course"><a href="<?php echo admin_url('admin.php?page=course_details&course_id=' . $course_object->ID); ?>"><?php _e('Edit', 'cp'); ?></a> | </span>
+                                        <?php if (current_user_can('coursepress_view_all_units_cap') || $course_object->post_author == get_current_user_id()) { ?>
+                                            <span class="course_units"><a href="<?php echo admin_url('admin.php?page=course_details&tab=units&course_id=' . $course_object->ID); ?>"><?php _e('Units', 'cp'); ?></a> | </span>
                                         <?php } ?>
-                                        <span class="course_students"><a href="?page=course_details&tab=students&course_id=<?php echo $course_object->ID; ?>"><?php _e('Students', 'cp'); ?></a> | </span>
+                                        <span class="course_students"><a href="<?php echo admin_url('admin.php?page=course_details&tab=students&course_id=' . $course_object->ID); ?>"><?php _e('Students', 'cp'); ?></a> | </span>
                                         <?php if (current_user_can('coursepress_change_course_status_cap') || (current_user_can('coursepress_change_my_course_status_cap') && $course_object->post_author == get_current_user_id())) { ?>
-                                            <span class="course_publish_unpublish"><a href="?page=courses&course_id=<?php echo $course_object->ID; ?>&action=change_status&new_status=<?php echo ($course_object->post_status == 'unpublished') ? 'publish' : 'private'; ?>"><?php ($course_object->post_status == 'unpublished') ? _e('Publish', 'cp') : _e('Unpublish', 'cp'); ?></a> | </span>
+                                            <span class="course_publish_unpublish"><a href="<?php echo wp_nonce_url(admin_url('admin.php?page=courses&course_id=' . $course_object->ID . '&action=change_status&new_status=' . ($course_object->post_status == 'unpublished' ? 'publish' : 'private')), 'change_course_status', 'cp_nonce'); ?>"><?php ($course_object->post_status == 'unpublished') ? _e('Publish', 'cp') : _e('Unpublish', 'cp'); ?></a> | </span>
                                         <?php } ?>
                                         <?php if (current_user_can('coursepress_delete_course_cap') || (current_user_can('coursepress_delete_my_course_cap') && $course_object->post_author == get_current_user_id())) { ?>
-                                            <span class="course_remove"><a href="?page=courses&action=delete&course_id=<?php echo $course_object->ID; ?>" onClick="return removeCourse();"><?php _e('Delete', 'cp'); ?></a> | </span>
+                                            <span class="course_remove"><a href="<?php echo wp_nonce_url(admin_url('admin.php?page=courses&action=delete&course_id=' . $course_object->ID), 'delete_course', 'cp_nonce'); ?>" onClick="return removeCourse();"><?php _e('Delete', 'cp'); ?></a> | </span>
                                         <?php } ?>
                                         <span class="view_course"><a href="<?php echo get_permalink($course->ID); ?>" rel="permalink"><?php _e('View Course', 'cp') ?></a><?php if (current_user_can('coursepress_view_all_units_cap') || $course_object->post_author == get_current_user_id()) { ?> | <?php } ?></span>
                                         <?php if (current_user_can('coursepress_view_all_units_cap') || $course_object->post_author == get_current_user_id()) { ?>
@@ -220,24 +225,24 @@ if (isset($_GET['quick_setup'])) {
                                     </div>
                                 </td>
                                 <td class="column-units <?php echo $style; ?>">
-                                <?php echo $course_obj->get_units('', 'any', true);?> <?php _e('Units', 'cp');?><br />
-                                <?php echo $course_obj->get_units('', 'publish', true);?> <?php _e('Published', 'cp');?>
+                                    <?php echo $course_obj->get_units('', 'any', true); ?> <?php _e('Units', 'cp'); ?><br />
+                                    <?php echo $course_obj->get_units('', 'publish', true); ?> <?php _e('Published', 'cp'); ?>
                                 </td>
-                                <td class="center column-students <?php echo $style; ?>"><a href="?page=course_details&tab=students&course_id=<?php echo $course_object->ID; ?>"><?php echo $course_obj->get_number_of_students(); ?></a></td>
+                                <td class="center column-students <?php echo $style; ?>"><a href="<?php echo admin_url('admin.php?page=course_details&tab=students&course_id=' . $course_object->ID); ?>"><?php echo $course_obj->get_number_of_students(); ?></a></td>
                                 <td class="column-status <?php echo $style; ?>"><?php echo ($course_object->post_status == 'publish') ? ucfirst($course_object->post_status) . 'ed' : ucfirst($course_object->post_status); ?></td>
                                 <td class="column-actions <?php echo $style; ?>">
-                                    <a href="?page=course_details&course_id=<?php echo $course_object->ID; ?>" class="button button-settings"><?php _e('Settings', 'cp'); ?></a>
+                                    <a href="<?php echo admin_url('admin.php?page=course_details&course_id=' . $course_object->ID); ?>" class="button button-settings"><?php _e('Settings', 'cp'); ?></a>
 
                                     <?php if (current_user_can('coursepress_view_all_units_cap') || $course_object->post_author == get_current_user_id()) { ?>
-                                        <a href="?page=course_details&tab=units&course_id=<?php echo $course_object->ID; ?>" class="button button-units"><?php _e('Units', 'cp'); ?></a>
+                                        <a href="<?php echo admin_url('admin.php?page=course_details&tab=units&course_id=' . $course_object->ID); ?>" class="button button-units"><?php _e('Units', 'cp'); ?></a>
                                     <?php } ?>
                                     <?php if (current_user_can('coursepress_change_course_status_cap') || (current_user_can('coursepress_change_my_course_status_cap') && $course_object->post_author == get_current_user_id())) { ?>
-                                        <a href="?page=courses&course_id=<?php echo $course_object->ID; ?>&action=change_status&new_status=<?php echo ($course_object->post_status == 'unpublished') ? 'publish' : 'private'; ?>" class="button button-<?php echo ($course_object->post_status == 'unpublished') ? 'publish' : 'unpublish'; ?>"><?php ($course_object->post_status == 'unpublished') ? _e('Publish', 'cp') : _e('Unpublish', 'cp'); ?></a></td>
+                                        <a href="<?php echo wp_nonce_url(admin_url('admin.php?page=courses&course_id=' . $course_object->ID . '&action=change_status&new_status=' . ($course_object->post_status == 'unpublished' ? 'publish' : 'private')), 'change_course_status', 'cp_nonce'); ?>" class="button button-<?php echo ($course_object->post_status == 'unpublished') ? 'publish' : 'unpublish'; ?>"><?php ($course_object->post_status == 'unpublished') ? _e('Publish', 'cp') : _e('Unpublish', 'cp'); ?></a></td>
                                 <?php } ?>
                                 <?php if (current_user_can('coursepress_delete_course_cap') || (current_user_can('coursepress_delete_my_course_cap'))) { ?>
                                     <td class="column-remove <?php echo $style; ?>">
                                         <?php if (current_user_can('coursepress_delete_course_cap') || (current_user_can('coursepress_delete_my_course_cap') && $course_object->post_author == get_current_user_id())) { ?>
-                                            <a href="?page=courses&action=delete&course_id=<?php echo $course_object->ID; ?>" onClick="return removeCourse();">
+                                            <a href="<?php echo wp_nonce_url(admin_url('admin.php?page=courses&action=delete&course_id=' . $course_object->ID), 'delete_course', 'cp_nonce'); ?>" onClick="return removeCourse();">
                                                 <i class="fa fa-times-circle cp-move-icon remove-btn"></i>
                                             </a>
                                         <?php } ?>
